Adding all old changes

+++ Services/RestaurantService.cs
@@ -1,12 +1,15 @@
 ﻿using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
+using RestaurantAPI.Authorization;
 using RestaurantAPI.Entities;
 using RestaurantAPI.Exceptions;
 using RestaurantAPI.Services.Abstract;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace RestaurantAPI.Services
@@ -17,16 +20,17 @@ namespace RestaurantAPI.Services
 		private readonly RestaurantDbContext _dbContext;
 		private readonly IMapper _mapper;
 		private readonly ILogger <RestaurantService> _logger;
+		private readonly IAuthorizationService _authorizationService;
 
-
-		public RestaurantService(RestaurantDbContext dbContext,IMapper mapper,ILogger<RestaurantService> logger)
+		public RestaurantService(RestaurantDbContext dbContext,IMapper mapper,ILogger<RestaurantService> logger,IAuthorizationService authorizationService)
 		{
 			_dbContext = dbContext;
 			_mapper = mapper;
 			_logger = logger;
+			_authorizationService = authorizationService;
 		}
 
-		public void Delete (int id)
+		public void Delete (int id, ClaimsPrincipal user)
 		{
 			_logger.LogWarning($"Restaurant with id: {id} DELETE action invoked");
 			var restaurant = _dbContext
@@ -35,7 +39,14 @@ namespace RestaurantAPI.Services
 
 			if (restaurant is null)
 				throw new NotFoundException("Restaurant not found");
-			
+
+			var authorizationResult = _authorizationService.AuthorizeAsync(user, restaurant, new ResourcesOperationRequirement(ResourceOperation.Delete)).Result;
+
+			if (!authorizationResult.Succeeded)
+			{
+				throw new ForbidException();
+			}
+
 			_dbContext.Restaurant.Remove(restaurant);
 			_dbContext.SaveChanges();
 		
@@ -67,21 +78,32 @@ namespace RestaurantAPI.Services
 			var result = _mapper.Map<List<RestaurantDto>>(restaurants);
 			return result;
 		}
-		public int Create(CreateRestaurantDto dto)
+		public int Create(CreateRestaurantDto dto,int userId)
 		{
 			var restaurant = _mapper.Map<Restaurant>(dto);
+			restaurant.CreateById = userId;
 			_dbContext.Restaurant.Add(restaurant);
 			_dbContext.SaveChanges();
 			return restaurant.Id;
 		}
 
-		public void Update(int id,UpdateRestaurantDto dto)
+		public void Update(int id,UpdateRestaurantDto dto,ClaimsPrincipal user)
 		{
+			
 			var restById = _dbContext
 				   .Restaurant
 				   .FirstOrDefault(r => r.Id == id);
+
+			
 			if (restById is null)
 				throw new NotFoundException("Restaurant not found");
+
+			var authorizationResult=_authorizationService.AuthorizeAsync(user, restById, new ResourcesOperationRequirement(ResourceOperation.Update)).Result;
+
+			if (!authorizationResult.Succeeded)
+			{
+				throw new ForbidException();
+			}
 
 			restById.Name = dto.Name;
 			restById.Description = dto.Description;
